requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25

# HTTP Client
requests==2.31.0

# Hardware - NVIDIA Jetson Nano GPIO
# Installed separately by install.sh (requires system-level Jetson libs)
# Jetson.GPIO==2.1.6

# Configuration
pyyaml==6.0.1
